tests/wpt/tests/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-case-xhtml.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>getElementsByName and case</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-getelementsbyname"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<div id="test">
<div name="abcd"></div>
</div>
<script>
test(function() {
  assert_equals(document.getElementsByName("ABCD").length, 0);
  assert_equals(document.getElementsByName("abcd").length, 1);
});
</script>
</body>
</html>
